requirements.txt
fast-bitrix24==1.5.11
html2text==2020.1.16
imap-tools==0.55.0
peewee==3.14.10
python-dateutil==2.8.2
python-dotenv==0.20.0
python-telegram-bot==13.12
urllib3==1.26.9
